pyproject.toml
[project]
name = "project-rag-klasifikasi-gambar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3",
]
